\name{sangerseqR-package}
\alias{sangerseqR-package}
\alias{sangerseqR}
\docType{package}
\title{
Tools for Sanger Sequencing Data in R
}
\description{
This package contains several tools for analyzing Sanger Sequencing data files 
in R, including reading .scf and .ab1 files, making basecalls and plotting 
chromatograms.
}
\author{
Jonathon T. Hill, Bradley Demarest

Maintainer: Jonathon Hill <jhill@genetics.utah.edu>
}
\references{
  Jonathon T Hill, Bradley L Demarest, Brent W Bisgrove, Yi-chu Su, Megan Smith and H. Joseph
  Yost (2014). Poly peak parser: Method and software for identification of unknown indels
  using sanger sequencing of polymerase chain reaction products. Developmental Dynamics
  243:1632-1636
}
\seealso{
Biostrings
}
