% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PPPfunction.R
\name{PolyPeakParser}
\alias{PolyPeakParser}
\title{Run Poly Peak Parser}
\usage{
PolyPeakParser()
}
\value{
\code{\link{scf}} s4 object
}
\description{
Runs the Poly Peak Parser shiny (shiny.rstudio.com) app in the system's default browser. Poly Peak Parser
is a web front end that reads, plots and parses double peaks from chromatogram files.
Instructions can be found on the webpage once it launches.
}
\examples{
\dontrun{
PolyPeakParser()
}
}
\seealso{
\code{\link{read.abif}}, \code{\link{readsangerseq}},  \code{\link{scf}}
}

