\name{sangerseqR-package}
\alias{sangerseqR-package}
\alias{sangerseqR}
\docType{package}
\title{
Tools for Sanger Sequencing Data in R
}
\description{
This package contains several tools for analyzing Sanger Sequencing data files 
in R, including reading .scf and .ab1 files, making basecalls and plotting 
chromatograms.
}
\author{
Jonathon T. Hill, Bradley Demarest

Maintainer: Jonathon Hill <jhill@genetics.utah.edu>
}
\references{
Hill et al. (submitted)
}
\seealso{
Biostrings
}
